넉백 개선 및 버그 수정

+++ Assets/Scripts/Data/CharacterAtkGameData.cs
@@ -41,7 +41,7 @@ namespace QT
         
         public float TeleportAllowableDistance { get; set; }
         public float ReflectCorrection { get; set; }
-        public int AtkDmgPer { get; set; }
+        public float AtkDmgPer { get; set; }
     }
 
     [GameDataBase((typeof(CharacterAtkGameData)),"CharacterAtkGameData")]
